Converted rawTransactionTests to example class
The rawTransactionTests were relatively instable (successful the first
3-4-5 times they were ran on regtest, but failing afterwards).

Reason for this was the hardcoded transaction fee in the test, since
there doesn't seem to be a reliable way to get estimated tx fees on
regtest.

So to avoid having these test scenarios constantly fail the test suite,
they have instead been converted and moved to an examples package.